Rename client-package to a more descriptive name
The base package of this project was called client. This is a very
generic, meaningless name, especially if we are to publish it as a
package on pypi for instance.

This is a first step in separating import code from dataset specific
information.
