+++ src/Dockerfile
@@ -1,7 +1,7 @@
 FROM amsterdam/docker_python:latest
 MAINTAINER datapunt@amsterdam.nl
 
-# Install client in /app folder
+# Install gobimportclient in /app folder
 WORKDIR /app
 
 # Install required Python packages
@@ -9,5 +9,5 @@ COPY requirements.txt /app/
 RUN pip3 install --no-cache-dir -r requirements.txt
 RUN rm requirements.txt
 
-# Copy client module
-COPY client client
+# Copy gobimportclient module
+COPY gobimportclient gobimportclient
